Move TestCore* files for core folder
